Update to rustc 1.3.0-dev (4b4119d5c 2015-07-29)

+++ src/data/achievements.rs
@@ -1,12 +1,12 @@
 use std::collections::HashMap;
 
-#[deriving(Show, Encodable, Decodable)]
+#[derive(Debug, RustcEncodable, RustcDecodable)]
 pub struct Achievements {
     pub beastMaster: bool,
     //challenges: Vec<?>,
-    pub perfect: uint,
-    pub quests: HashMap<String, uint>,
-    pub streak: uint,
+    pub perfect: usize,
+    pub quests: HashMap<String, usize>,
+    pub streak: usize,
     pub ultimateGear: bool,
 }
